# [Performance] Support MQA/GQA in decode phase by using FlashAttention