--- a/csrc/cache.h
+++ b/csrc/cache.h
@@ -15,6 +15,13 @@ void copy_blocks(
   std::vector<torch::Tensor>& value_caches,
   const std::map<int64_t, std::vector<int64_t>>& block_mapping);
 
+void cache(
+  torch::Tensor& key,
+  c10::optional<torch::Tensor>& value,
+  torch::Tensor& key_cache,
+  c10::optional<torch::Tensor>& value_cache,
+  torch::Tensor& slot_mapping);
+
 void reshape_and_cache(
   torch::Tensor& key,
   torch::Tensor& value,