--- a/vllm/config.py
+++ b/vllm/config.py
@@ -55,6 +55,8 @@ class ModelConfig:
         max_context_len_to_capture: Maximum context len covered by CUDA graphs.
             When a sequence has context length larger than this, we fall back
             to eager mode.
+        use_flash_attn_zte: Use original flash attention for prefill stage and
+            use blocked kv cache flash attention for decode stage.
     """
 
     def __init__(
@@ -73,6 +75,7 @@ class ModelConfig:
         quantization: Optional[str] = None,
         enforce_eager: bool = False,
         max_context_len_to_capture: Optional[int] = None,
+        use_flash_attn_zte: bool = False,
     ) -> None:
         self.model = model
         self.tokenizer = tokenizer
@@ -86,6 +89,7 @@ class ModelConfig:
         self.quantization = quantization
         self.enforce_eager = enforce_eager
         self.max_context_len_to_capture = max_context_len_to_capture
+        self.use_flash_attn_zte = use_flash_attn_zte
 
         if os.environ.get("VLLM_USE_MODELSCOPE", "False").lower() == "true":
             # download model from ModelScope hub,