--- a/vllm/config.py
+++ b/vllm/config.py
@@ -55,8 +55,7 @@ class ModelConfig:
         max_context_len_to_capture: Maximum context len covered by CUDA graphs.
             When a sequence has context length larger than this, we fall back
             to eager mode.
-        use_flash_attn_zte: Use original flash attention for prefill stage and
-            use blocked kv cache flash attention for decode stage.
+        use_flash_attn: Use blocked kv cache flash attention for decode stage.
     """
 
     def __init__(
@@ -75,7 +74,7 @@ class ModelConfig:
         quantization: Optional[str] = None,
         enforce_eager: bool = False,
         max_context_len_to_capture: Optional[int] = None,
-        use_flash_attn_zte: bool = False,
+        use_flash_attn: bool = False,
     ) -> None:
         self.model = model
         self.tokenizer = tokenizer
@@ -89,7 +88,7 @@ class ModelConfig:
         self.quantization = quantization
         self.enforce_eager = enforce_eager
         self.max_context_len_to_capture = max_context_len_to_capture
-        self.use_flash_attn_zte = use_flash_attn_zte
+        self.use_flash_attn = use_flash_attn
 
         if os.environ.get("VLLM_USE_MODELSCOPE", "False").lower() == "true":
             # download model from ModelScope hub,