--- a/vllm/model_executor/layers/attention.py
+++ b/vllm/model_executor/layers/attention.py
@@ -3,6 +3,7 @@ from typing import List, Optional
 
 import torch
 import torch.nn as nn
+from flash_attn.flash_attn_interface import flash_attn_with_blocked_kvcache
 from xformers import ops as xops
 from xformers.ops.fmha.attn_bias import (BlockDiagonalCausalMask,
                                          LowerTriangularMaskWithTensorBias)
@@ -71,10 +72,10 @@ class PagedAttention(nn.Module):
             query: shape = [batch_size, seq_len, num_heads * head_size]
             key: shape = [batch_size, seq_len, num_kv_heads * head_size]
             value: shape = [batch_size, seq_len, num_kv_heads * head_size]
-            key_cache: shape = [num_blocks, num_kv_heads, head_size/x,
-                block_size, x]
-            value_cache: shape = [num_blocks, num_kv_heads, head_size,
-                block_size]
+            key_cache: shape = [num_blocks, block_size, num_kv_heads,
+                head_size]
+            value_cache: shape = [num_blocks, block_size, num_kv_heads,
+                head_size]
             input_metadata: metadata for the inputs.
         Returns:
             shape = [batch_size, seq_len, num_heads * head_size]
@@ -90,7 +91,7 @@ class PagedAttention(nn.Module):
         # vectors will not be cached. This happens during the initial memory
         # profiling run.
         if key_cache is not None and value_cache is not None:
-            cache_ops.reshape_and_cache(
+            cache_ops.cache(
                 key,
                 value,
                 key_cache,
@@ -156,14 +157,15 @@ class PagedAttention(nn.Module):
             output = out.view_as(query)
         else:
             # Decoding run.
-            output = _paged_attention(
-                query,
+            output = flash_attn_with_blocked_kvcache(
+                query.unsqueeze(1),
                 key_cache,
                 value_cache,
-                input_metadata,
-                self.num_kv_heads,
-                self.scale,
-                self.alibi_slopes,
+                input_metadata.block_tables,
+                input_metadata.context_lens,
+                softmax_scale=self.scale,
+                causal=True,
+                alibi_slopes=self.alibi_slopes,
             )
 
         # Reshape the output tensor.