--- a/vllm/worker/cache_engine.py
+++ b/vllm/worker/cache_engine.py
@@ -51,20 +51,17 @@ class CacheEngine:
         self.events = [torch.cuda.Event() for _ in range(self.num_layers)]
 
     def get_key_block_shape(self) -> Tuple[int, int, int, int]:
-        element_size = torch.tensor([], dtype=self.dtype).element_size()
-        x = 16 // element_size
         return (
-            self.num_heads,
-            self.head_size // x,
             self.block_size,
-            x,
+            self.num_heads,
+            self.head_size,
         )
 
     def get_value_block_shape(self) -> Tuple[int, int, int]:
         return (
+            self.block_size,
             self.num_heads,
             self.head_size,
-            self.block_size,
         )
 
     def allocate_gpu_cache(self) -> List[KVCache]: