As shown in issue #1880, vLLM's current paged attention kernel does not leverage the benefits of MQA/GQA. FlashAttention supports MQA/GQA at kernel level and has supported paged kv cache since v2.5.0. 

To enjoy the benefit of MQA/GQA, this PR replaces `_paged_attention` with `flash_attn_with_kvcache`.

Notes:
- As paged KV cache block size in Flash Attention must be divisible by 256, the block size is set to 256 by default in https://github.com/vllm-project/vllm/pull/2744/files#diff-ea8b8ff63961713ccb62d78e53e96404b587b7828cb9fee08a9e5576bf563673R54 So GPU Blocks will decrease significantly compared to before. It is not a bug. But this will waste gpu memory especially when batch size is large. And also `cache_ops.cache` will cost more time than original version (block_size=16). **So this PR is mainly for small batch size cases.**
- This PR is mainly for MQA/GQA model. MHA models will not gain much speedup based on this PR.

# Kernel latency
Env: 
- A100-40GB-PCIE
- num query heads=64, num kv heads=8
- kernel data type=bfloat16
- use_alibi=False

Below are results by running `benchmark_paged_attention.py`.
context_len=1024 : 
Batchsize | Paged attention V1 (us) | Paged attention V2 (us) |  FA (us) | Speedup (min(V1, V2) / FA)
---|---|---|---|---
1 | 37.541 | 29.602 | 27.675 | 1.07
4 | 68.585 | 71.447 | 27.433 | 2.50
16 | 162.237 | 168.997 |  67.681 | 2.40
64 | 687.938 | 733.687 | 214.041 | 3.21
256 | 2215.532 | 2378.606 | 680.123 | 3.26

context_len=4096 : 
Batchsize | Paged attention V1 (us) | Paged attention V2 (us) |  FA (us) | Speedup (min(V1, V2) / FA)
---|---|---|---|---
1 | 137.858 | 70.783 | 28.557 | 2.48
4 | 223.438 | 164.293 | 62.104 | 2.65
16 | 716.780 | 724.601 |  238.730 | 3.00
64 | 2324.863 | 2792.657 | 772.266 | 3.01
256 | 8771.168 | 9723.521 | 2583.970 | 3.39

# E2E latency

# E2E throughput
Env: 
- 4xA100-40GB-PCIE
- Baseline: `python benchmarks/benchmark_throughput.py --input-len 512 --output-len 512 --model /CodeLlama-34b-hf/ --tokenizer /CodeLlama-34b-hf/ --trust-remote-code --tensor-parallel-size 4 --enforce-eager`
- FA: `python benchmarks/benchmark_throughput.py --input-len 512 --output-len 512 --model /CodeLlama-34b-hf/ --tokenizer /CodeLlama-34b-hf/ --trust-remote-code --tensor-parallel-size 4 --enforce-eager --use-flash-attn`

Below are results by running `benchmark_throughput.py`.
Original (tokens/sec) | FA (tokens/sec) | Speedup (FA / Original)
---|---|---
2114.28 | 2311.75 | 1.09