@ttbachyinsda 2024-02-05
It appears that the throughput test for qwen-7b has a negative improvement, and it seems that the GPU blocks have decreased significantly compared to before. Maybe there is a bug?

No Flash Attention:
Namespace(backend='vllm', dataset=None, input_len=128, output_len=128, model='./Qwen-7B-Chat', tokenizer='./Qwen-7B-Chat', quantization=None, tensor_parallel_size=1, n=1, use_beam_search=False, num_prompts=1000, seed=0, hf_max_batch_size=None, trust_remote_code=True, max_model_len=None, dtype='auto', enforce_eager=False, kv_cache_dtype='auto', use_flash_attn=False, device='cuda')INFO 02-05 19:24:49 llm_engine.py:73] Initializing an LLM engine with config: model='./Qwen-7B-Chat', tokenizer='./Qwen-7B-Chat', tokenizer_mode=auto, revision=None, tokenizer_revision=None, trust_remote_code=True, dtype=torch.float16, max_seq_len=8192, download_dir=None, load_format=auto, tensor_parallel_size=1, disable_custom_all_reduce=False, quantization=None, enforce_eager=False, kv_cache_dtype=auto, use_flash_attn=False, device_config=cuda, seed=0)                                                                                                                                                                                                             WARNING 02-05 19:24:50 tokenizer.py:64] Using a slow tokenizer. This might cause a significant slowdown. Consider using a fast tokenizer instead.INFO 02-05 19:25:21 llm_engine.py:331] # GPU blocks: 643, # CPU blocks: 512                                                                                                                                                                              INFO 02-05 19:25:22 model_runner.py:653] Capturing the model for CUDA graphs. This may lead to unexpected consequences if the model is not static. To run the model in eager mode, set 'enforce_eager=True' or use '--enforce-eager' in the CLI.INFO 02-05 19:25:22 model_runner.py:657] CUDA graphs can take additional 1~3 GiB memory per GPU. If you are running out of memory, consider decreasing `gpu_memory_utilization` or enforcing eager mode. You can also reduce the `max_num_seqs` as needed to decrease memory usage.                                                                                                                                                                                                                               INFO 02-05 19:25:26 model_runner.py:720] Graph capturing finished in 4 secs.                                                                                                                                                                             Processed prompts: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1000/1000 [02:21<00:00,  7.07it/s]Throughput: 7.07 requests/s, 1810.59 tokens/s

With Flash Attention:
Namespace(backend='vllm', dataset=None, input_len=128, output_len=128, model='./Qwen-7B-Chat', tokenizer='./Qwen-7B-Chat', quantization=None, tensor_parallel_size=1, n=1, use_beam_search=False, num_prompts=1000, seed=0, hf_max_batch_size=None, trust_remote_code=True, max_model_len=None, dtype='auto', enforce_eager=False, kv_cache_dtype='auto', use_flash_attn=True, device='cuda')INFO 02-05 19:31:59 llm_engine.py:73] Initializing an LLM engine with config: model='./Qwen-7B-Chat', tokenizer='./Qwen-7B-Chat', tokenizer_mode=auto, revision=None, tokenizer_revision=None, trust_remote_code=True, dtype=torch.float16, max_seq_len=8192, download_dir=None, load_format=auto, tensor_parallel_size=1, disable_custom_all_reduce=False, quantization=None, enforce_eager=False, kv_cache_dtype=auto, use_flash_attn=True, device_config=cuda, seed=0)WARNING 02-05 19:32:00 tokenizer.py:64] Using a slow tokenizer. This might cause a significant slowdown. Consider using a fast tokenizer instead.INFO 02-05 19:32:14 llm_engine.py:331] # GPU blocks: 40, # CPU blocks: 32INFO 02-05 19:32:15 model_runner.py:653] Capturing the model for CUDA graphs. This may lead to unexpected consequences if the model is not static. To run the model in eager mode, set 'enforce_eager=True' or use '--enforce-eager' in the CLI.INFO 02-05 19:32:15 model_runner.py:657] CUDA graphs can take additional 1~3 GiB memory per GPU. If you are running out of memory, consider decreasing `gpu_memory_utilization` or enforcing eager mode. You can also reduce the `max_num_seqs` as needed to decrease memory usage.INFO 02-05 19:32:19 model_runner.py:720] Graph capturing finished in 5 secs.Processed prompts: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1000/1000 [05:09<00:00,  3.23it/s]Throughput: 3.23 requests/s, 826.13 tokens/s