@casper-hansen 2024-02-05
@zhaoyang-star you may be interested in FlashInfer's PagedAttention which is 3x faster than vLLM's version. It supports GQA

https://github.com/flashinfer-ai/flashinfer/
https://flashinfer.ai/2024/02/02/introduce-flashinfer.html

![image](https://github.com/vllm-project/vllm/assets/27340033/218fd1ee-bc64-4b47-b10b-5d4f0a22da24)